+++ test/DotNetWorkerTests/GrpcWorkerTests.cs
@@ -205,8 +205,25 @@ namespace Microsoft.Azure.Functions.Worker.Tests
             var response = await invocationHandler.InvokeAsync(request);
 
             Assert.Equal(StatusResult.Types.Status.Failure, response.Result.Status);
-            Assert.Contains("InvalidOperationException: whoops", response.Result.Exception.Message);
-            Assert.Contains("CreateContext", response.Result.Exception.Message);
+            Assert.Contains("whoops", response.Result.Exception.Message);
+            Assert.Contains("CreateContext", response.Result.Exception.StackTrace);
+        }
+
+        [Fact]
+        public async Task Invoke_InvokeAsyncThrows_ReturnsFailure()
+        {
+            _mockApplication
+                .Setup(m => m.InvokeFunctionAsync(It.IsAny<FunctionContext>()))
+                .Throws(new InvalidOperationException("whoops"));
+
+            var request = CreateInvocationRequest();
+
+            var response = await GrpcWorker.InvocationRequestHandlerAsync(request, _mockApplication.Object, _mockFeaturesFactory.Object,
+                new JsonObjectSerializer(), _mockOutputBindingsInfoProvider.Object, _mockInputConversionFeatureProvider.Object);
+
+            Assert.Equal(StatusResult.Types.Status.Failure, response.Result.Status);
+            Assert.Contains("whoops", response.Result.Exception.Message);
+            Assert.Contains("InvokeFunctionAsync", response.Result.Exception.StackTrace);
         }
 
         private static FunctionLoadRequest CreateFunctionLoadRequest()