+++ release_notes.md
@@ -3,7 +3,10 @@
 - My change description (#PR/#issue)
 -->
 
+- Added support for surfacing user-thrown exception to App Insights (#939)
 - Add Retry Policy Attributes (#977, #971)
 - Bug fix - GetOutputBindings returns incorrect data when OutputBindingData is not set (#983)
 - Worker version and environment information returned in init call (#1022)
 - Handle InvocationCancel message and signal cancellation for in-flight invocations (#972)
+
+- Added support for surfacing user-thrown exception to App Insights (#939)